Refactor ProjectCard component for improved structure and documentation
- Deleted the old `project-card.tsx` file and replaced it with a new implementation that enhances clarity and organization.
- Introduced a separate `project-card-props.ts` file to define the `ProjectCardProps` interface, improving type safety and documentation.
- Updated JSDoc comments in the `ProjectCard` component to provide detailed descriptions of props and functionality, enhancing maintainability.
